updated data 2/8/21

+++ application/views/fives/standardize/idea/index.php
@@ -4,14 +4,14 @@
 <?php if($this->session->flashdata('error_msg')) : ?>
     <p class="alert alert-dismissable alert-danger"><?php echo $this->session->flashdata('error_msg'); ?></p>
 <?php endif; ?>
-<div class="card-header"><h4>5S IDEA LIST
-    <a href="<?php echo base_url(); ?>fives/idea_add" class="btn btn-info float-right" style="margin-right:10px;">ADD</a>
+<div class="card-header" style="background-color:#1C4670; color:white;"><h4>5S IDEA LIST
+    <a href="<?php echo base_url(); ?>fives/idea_add" class="btn btn-dark float-right" style="border:1px solid #ccc; margin-right:10px;">ADD</a>
     </h4> 
 </div>
 <br>
 <table id="" class="display" style="width:100%">
     <thead>
-        <tr>
+        <tr style="background-color:#D4F1F4;">
             <th scope="col">Date</th>
             <th scope="col">Control No.</th>
             <th scope="col">Submitted By</th>
@@ -32,25 +32,35 @@
                     <td data-label="Position"><?php echo $idea->company; ?></td>
                     <td data-label="Date Hired"><?php echo $idea->department; ?></td>
                     <td data-label="Employee Status"><?php echo $idea->proposal; ?></td>
-                    <td data-label="Employee Status" <?php echo $idea->status == "Implemented" ? 'style="background-color:green; color:white;"' : ''; ?>><?php echo $idea->status; ?></td>
+                    <?php if($idea->status == "Open" ): ?>
+                        <td data-label="Employee Status" style="background-color:#A8D9F8; "><?php echo $idea->status; ?></td>
+                    <?php endif; ?>
+                    <?php if($idea->status == "Ongoing" ): ?>
+                        <td data-label="Employee Status" style="background-color:#FEDE00; "><?php echo $idea->status; ?></td>
+                    <?php endif; ?>
+                    <?php if($idea->status == "Implemented" ): ?>
+                        <td data-label="Employee Status" style="background-color:#7CF3A0; "><?php echo $idea->status; ?></td>
+                    <?php endif; ?>
                     <td data-label="Action">
                         <div class="btn-group">
                             <button type="button" class="btn btn-info dropdown-toggle btn-sm btnaction" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                                 Action
                             </button>
                             <div class="dropdown-menu dropdown-menu-right">
-                                <a class="dropdown-item" href="<?php echo base_url(); ?>fives/idea_view/<?php echo $idea->id; ?>"> View</a>
-                                <a class="dropdown-item" href="<?php echo base_url(); ?>fives/idea_edit/<?php echo $idea->id; ?>/<?php echo $idea->control_number; ?>">Edit</a>
-                                <a onclick="return confirm('Are you sure you want to delete data?');" class="dropdown-item" href="<?php echo base_url(); ?>fives/idea_delete/<?php echo $idea->id?>">Delete</a>
+                                <a class="dropdown-item" href="<?php echo base_url(); ?>fives/idea_view/<?php echo $idea->id; ?>/<?php echo $idea->control_number; ?>"> View</a>
+                                <?php if($this->session->userdata('username') == $idea->submit_by) : ?>
+                                    <a class="dropdown-item" href="<?php echo base_url(); ?>fives/idea_edit/<?php echo $idea->id; ?>/<?php echo $idea->control_number; ?>">Edit</a>
+                                    <a onclick="return confirm('Are you sure you want to delete data?');" class="dropdown-item" href="<?php echo base_url(); ?>fives/idea_delete/<?php echo $idea->id?>">Delete</a>
+                                <?php endif; ?>       
                                 <?php if($this->session->userdata('access_level_id') == 1 && $idea->status != "Implemented") : ?>
                                     <div class="dropdown-divider"></div>
                                     <a class="dropdown-item" href="<?php echo base_url(); ?>fives/status/<?php echo $idea->id; ?>/<?php echo $idea->control_number; ?>/<?php echo $idea->status; ?>">Status</a>
-                                    
                                 <?php endif; ?>    
                                 <?php if($this->session->userdata('access_level_id') == 1 && $idea->status != 'Open') : ?>
                                     <div class="dropdown-divider"></div>
                                     <a class="dropdown-item" href="<?php echo base_url(); ?>fives/edit_status/<?php echo $idea->control_number; ?>/<?php echo $idea->status; ?>">Edit Status</a>
                                 <?php endif; ?>  
+                               
                             </div>
                         </div>
                     </td>
@@ -103,11 +113,7 @@
                     extend: 'colvis',
                     text: 'Filter'
                 }
-            ],
-            columnDefs: [ {
-                targets: -1,
-                visible: false
-            } ]
+            ]
         } );
     } );
 </script>
